remove overlay from node dialog, remove usenoecontext, pass node from top of initialise discourse graph, create util
